site/requirements.txt
AMQPStorm==2.7.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
configparser==4.0.2
cryptography==2.8
globus-sdk==1.8.0
idna==2.8
multidict==4.6.1
mysql-connector-python==8.0.18
pamqp==2.3.0
protobuf==3.11.1
pycparser==2.19
PyJWT==1.7.1
requests==2.22.0
schedule==0.6.0
six==1.13.0
urllib3==1.25.7
yarl==1.3.0
